docs: add project logo
Generated with Nano Banana. I forgot the exact prompt but it was something like
this:

"Generate an image of a nixie, inside the NixOS logo with half the body behind
the logo, chin rest on hand, slightly smiling, cartoon style with large stroke,
used for a logo fits in square composition"

Took a few retries.